feat: order calculation and information pages
Calculations to add the number of items and the values have been included so that when you delete the item from the cart, the total value decreases.

Two new pages were also created. One to check the information of the selected orders, such as user information and payment method. However, the fields are not yet being validated as it is a basic application.

The other page shows the history of the order made with the date the request was made. When clicking on the icon the user is redirected to this page containing the history.